Refactors UI loading and updates theme
Removes direct UI element manipulation from the UI loading function for better separation of concerns.
The UI file is removed and will be generated at build time,
gitignore is updated to ignore the whole build folder, but keeps the debug folder and the .ui file.
Updates the default theme to a newer version.
Adds GTK_CSD and GTK_DEBUG environment variables to be set on program start.